Adding a separate submodule for _statstat functions
